+++ src/main/java/insane96mcp/enhancedai/modules/creeper/CreeperSwell.java
@@ -69,7 +69,7 @@ public class CreeperSwell extends Feature {
 	@Label(name = "Launch.inaccuracy", description = "The inaccuracy of the launching creeper in Normal difficulty, easy is increased, hard is decreased.")
 	public static Double launchInaccuracy = 0.5d;
 	@Config
-	@Label(name = "Launch.Explosion Radius", description = "The explosion radius of launching creepers. (Overwrites Cena creepers explosion radius)")
+	@Label(name = "Launch.Explosion Radius", description = "The explosion radius of launching creepers. Set to 0 to not change. (Overrides Cena creepers explosion radius)")
 	public static Integer launchExplosionRadius = 2;
 	@Config(min = 0d, max = 1d)
 	@Label(name = "Breach Chance", description = "Breaching creepers will try to open an hole in the wall to let mobs in.")
@@ -188,7 +188,8 @@ public class CreeperSwell extends Feature {
 
 		if (launch) {
 			creeper.goalSelector.addGoal(1, new AICreeperLaunchGoal(creeper));
-			compoundNBT.putByte("ExplosionRadius", launchExplosionRadius.byteValue());
+			if (launchExplosionRadius > 0)
+				compoundNBT.putByte("ExplosionRadius", launchExplosionRadius.byteValue());
 		}
 		creeper.readAdditionalSaveData(compoundNBT);
 		MessageCreeperDataSync.syncCreeperToPlayers(creeper);